# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PDF Translation & RAG API",
    "description": "PDF OCR, Translation, and Multimodal RAG services",
    "version": "2.1.0"
  },
  "paths": {
...
    "/multimodal/extract": {
      "post": {
        "tags": [
          "Multimodal Research"
        ],
        "summary": "Extract From Pdf Endpoint",
        "description": "Extracts text and visual elements from a PDF, summarizes images, and indexes content.\n\nPipeline:\n1. Phase 1: Extract text chunks and visual elements (PaddleOCR)\n2. Phase 2: Summarize visual elements (Gemini)\n3. Phase 3: Index all content to FAISS\n\nArgs:\n    file: The uploaded PDF file.\n    user_id: Authenticated user ID (injected via dependency).\n\nReturns:\n    ExtractedDocument containing text chunks and visual elements with summaries.\n\nRaises:\n    HTTPException: 400 for invalid input, 500 for processing errors.",
        "operationId": "extract_from_pdf_endpoint_multimodal_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_from_pdf_endpoint_multimodal_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/multimodal/file/{doc_id}": {
      "delete": {
        "tags": [
          "Multimodal Research"
        ],
        "summary": "Delete Multimodal Document",
        "description": "Deletes a multimodal document and all associated files.\n\nCleanup steps:\n1. Remove from FAISS vector index (text chunks + visual summaries)\n2. Delete physical files (original PDF + cropped images)\n\nArgs:\n    doc_id: Document UUID.\n    user_id: Authenticated user ID.\n\nReturns:\n    Success status dict.",
        "operationId": "delete_multimodal_document_multimodal_file__doc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Multimodal Document Multimodal File  Doc Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/graph/status": {
      "get": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "取得圖譜狀態",
        "description": "取得使用者知識圖譜的狀態資訊，包括節點數、邊數、社群數等。",
        "operationId": "get_graph_status_graph_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/graph/data": {
      "get": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "取得視覺化資料",
        "description": "取得 react-force-graph 格式的圖譜資料，用於前端視覺化。",
        "operationId": "get_graph_visualization_data_graph_data_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphVisualizationData"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/graph/rebuild": {
      "post": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "重建圖譜",
        "description": "強制重建使用者的知識圖譜。會重新從所有文件中抽取實體與關係。",
        "operationId": "rebuild_graph_graph_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRebuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/graph/optimize": {
      "post": {
        "tags": [
          "Knowledge Graph"
        ],
        "summary": "優化圖譜",
        "description": "執行實體融合與社群重建，合併相似的實體節點。",
        "operationId": "optimize_graph_graph_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_extract_from_pdf_endpoint_multimodal_extract_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extract_from_pdf_endpoint_multimodal_extract_post"
      },
...
      "ExtractedDocument": {
        "properties": {
          "doc_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doc Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "text_chunks": {
            "items": {
              "$ref": "#/components/schemas/TextChunk"
            },
            "type": "array",
            "title": "Text Chunks"
          },
          "visual_elements": {
            "items": {
              "$ref": "#/components/schemas/VisualElement"
            },
            "type": "array",
            "title": "Visual Elements"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "user_id",
          "text_chunks",
          "visual_elements"
        ],
        "title": "ExtractedDocument"
      },
...
      "GraphOperationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "GraphOperationResponse",
        "description": "Response for graph operations."
      },
      "GraphOptimizeRequest": {
        "properties": {
          "regenerate_communities": {
            "type": "boolean",
            "title": "Regenerate Communities",
            "description": "重新生成社群摘要",
            "default": true
          }
        },
        "type": "object",
        "title": "GraphOptimizeRequest",
        "description": "Request for graph optimization."
      },
      "GraphRebuildRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "強制重建即使圖譜沒有變更",
            "default": false
          }
        },
        "type": "object",
        "title": "GraphRebuildRequest",
        "description": "Request for graph rebuild."
      },
      "GraphStatusResponse": {
        "properties": {
          "has_graph": {
            "type": "boolean",
            "title": "Has Graph",
            "description": "是否有圖譜"
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count",
            "description": "節點數量",
            "default": 0
          },
          "edge_count": {
            "type": "integer",
            "title": "Edge Count",
            "description": "邊數量",
            "default": 0
          },
          "community_count": {
            "type": "integer",
            "title": "Community Count",
            "description": "社群數量",
            "default": 0
          },
          "pending_resolution": {
            "type": "integer",
            "title": "Pending Resolution",
            "description": "待融合實體數量",
            "default": 0
          },
          "needs_optimization": {
            "type": "boolean",
            "title": "Needs Optimization",
            "description": "是否需要優化",
            "default": false
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated",
            "description": "最後更新時間"
          }
        },
        "type": "object",
        "required": [
          "has_graph"
        ],
        "title": "GraphStatusResponse",
        "description": "Response model for graph status API endpoint.\n\nProvides overview information about the user's knowledge graph.",
        "example": {
          "community_count": 12,
          "edge_count": 3400,
          "has_graph": true,
          "last_updated": "2025-12-21T10:30:00Z",
          "needs_optimization": true,
          "node_count": 1250,
          "pending_resolution": 45
        }
      },
      "GraphVisualizationData": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/VisNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/VisLink"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "type": "object",
        "title": "GraphVisualizationData",
        "description": "Response for graph visualization (react-force-graph format)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TextChunk": {
        "properties": {
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Markdown formatted content"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "page_number",
          "content",
          "chunk_id"
        ],
        "title": "TextChunk"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisLink": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "來源節點 ID"
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "目標節點 ID"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "關係標籤"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "label"
        ],
        "title": "VisLink",
        "description": "Link for react-force-graph visualization."
      },
      "VisNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "唯一節點識別碼（標籤）"
          },
          "group": {
            "type": "integer",
            "title": "Group",
            "description": "分群 ID（用於著色）"
          },
          "val": {
            "type": "integer",
            "title": "Val",
            "description": "節點大小（引用次數）"
          },
          "desc": {
            "type": "string",
            "title": "Desc",
            "description": "節點描述"
          }
        },
        "type": "object",
        "required": [
          "id",
          "group",
          "val",
          "desc"
        ],
        "title": "VisNode",
        "description": "Node for react-force-graph visualization."
      },
      "VisualElement": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/VisualElementType"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path",
            "description": "Path to the cropped image"
          },
          "bbox": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Bbox",
            "description": "[x1, y1, x2, y2] coordinates"
          },
          "original_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "Gemini generated summary"
          },
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text",
            "description": "周圍文字上下文（Caption + 前後文）"
          },
          "figure_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figure Reference",
            "description": "圖片引用標識，如 'Figure 1', '圖一'"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "page_number",
          "image_path",
          "bbox"
        ],
        "title": "VisualElement"
      },
      "VisualElementType": {
        "type": "string",
        "enum": [
          "table",
          "figure",
          "formula"
        ],
        "title": "VisualElementType"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}